clean up transaction highlighting logic and add explanation comments
